// src/types.ts
export type UserRole = 'SUPER_ADMIN' | 'ADMIN_TOKO' | 'KARYAWAN';
// ...
export interface ActivityLog {
  id: string;
  user_id: string;
  user_name: string;
  role?: UserRole;
  store_id: string | null;
  store_name?: string;
  aktivitas: string;
  detail_perubahan: string;
  waktu: string;
  ip_address?: string;
  tipe?: 'INVENTORY' | 'AUTH' | 'RETURN' | 'ABSENSI' | 'STORE' | 'USER';
}

export interface AppNotification {
  id: string;
  store_id?: string | null;
  store_name?: string;
  target_roles?: UserRole[];
  target_user_id?: string | null;
  creator_id?: string;
  creator_name?: string;
  kategori?: 'INVENTORY' | 'RETURN' | 'ABSENSI' | 'SISTEM' | 'AUTH';
  judul: string;
  pesan: string;
  tipe: 'WARNING' | 'SUCCESS' | 'INFO' | 'DANGER';
  waktu: string;
  read: boolean;
  read_by?: string[];
  link_page?: NavigationPage | string;
  is_global?: boolean;
  created_at?: string;
}

export type NavigationPage = 
  | 'PILIH_TOKO'
  | 'DASHBOARD_UTAMA'
  | 'DASHBOARD_TOKO'
  | 'INVENTARIS'
  | 'KATEGORI'
  | 'BARANG_MASUK'
  | 'BARANG_KELUAR'
  | 'BARANG_RUSAK'
  | 'RETURN'
  | 'ABSENSI'
  | 'KARYAWAN'
  | 'LAPORAN'
  | 'ACTIVITY_LOG'
  | 'TOKO'
  | 'PENGATURAN'
  | 'PROFIL';